requirement.txt
asgiref==3.5.2
autopep8==1.7.0
distlib==0.3.6
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
filelock==3.8.0
platformdirs==2.5.2
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.5
virtualenv==20.16.5
